purge: remove api/enums.py as it is unneeded (#1159)
* purge: remove api/enums.py as it is unneeded

* fix: remove bad import
