+++ interactions/api/__init__.py
@@ -5,7 +5,6 @@ This section of the library maintains and
 handles all the Gateway and HTTP work.
 """
 from .cache import *  # noqa: F401 F403
-from .enums import *  # noqa: F401 F403
 from .error import *  # noqa: F401 F403
 from .gateway import *  # noqa: F401 F403
 from .http import *  # noqa: F401 F403
