+++ interactions/api/gateway/client.py
@@ -16,6 +16,7 @@ from asyncio import (
     wait,
     wait_for,
 )
+from enum import IntEnum
 from sys import platform, version_info
 from time import perf_counter
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Type, Union
@@ -28,7 +29,6 @@ from ...client.enums import ComponentType, InteractionType, OptionType
 from ...client.models import Option
 from ...utils.missing import MISSING
 from ..dispatch import Listener
-from ..enums import OpCodeType
 from ..error import LibraryException
 from ..http.client import HTTPClient
 from ..models.flags import Intents
@@ -49,7 +49,31 @@ if TYPE_CHECKING:
 
 log = get_logger("gateway")
 
-__all__ = ("WebSocketClient",)
+__all__ = ("WebSocketClient", "OpCodeType")
+
+
+class OpCodeType(IntEnum):
+    """
+    An enumerable object for the Gateway's OPCODE result state.
+    This is representative of the OPCodes generated by the WebSocket.
+
+    .. note::
+        Equivalent of `Gateway Opcodes <https://discord.com/developers/docs/topics/opcodes-and-status-codes#gateway-opcodes>`_ in the Discord API.
+    """
+
+    DISPATCH = 0
+    HEARTBEAT = 1
+    IDENTIFY = 2
+    PRESENCE = 3
+    VOICE_STATE = 4
+    VOICE_PING = 5
+    RESUME = 6
+    RECONNECT = 7
+    REQUEST_MEMBERS = 8
+    INVALIDATE_SESSION = 9
+    HELLO = 10
+    HEARTBEAT_ACK = 11
+    GUILD_SYNC = 12
 
 
 class WebSocketClient:
